// chat2db-community-client/src/typings/sqlParser.ts
export interface ISqlEditorHintRangeVO {
  startLineNumber: number;
  startColumn: number;
  endLineNumber: number;
  endColumn: number;
}

export interface ISqlEditorHintItemVO {
  rowIndex: number;
  columnIndex: number;
  fieldName?: string;
  fieldType?: string;
  label?: string;
  range?: ISqlEditorHintRangeVO;
  active?: boolean;
}

export interface ISqlEditorHintVO {
  type: 'INSERT_VALUE' | 'ROUTINE_PARAMETER' | string;
  statementRange?: ISqlEditorHintRangeVO;
  rowRange?: ISqlEditorHintRangeVO;
  valueRange?: ISqlEditorHintRangeVO;
  items?: ISqlEditorHintItemVO[];
}